refactor(benchmark): move jmh packaging tasks to script
